+++ ui/panels-plugin/src/plugins/stat-chart/StatChartOptionsEditor.tsx
@@ -12,12 +12,42 @@
 // limitations under the License.
 
 import { JSONEditor } from '@perses-dev/components';
-import { OptionsEditorProps } from '@perses-dev/plugin-system';
+import {
+  OptionsEditorProps,
+  OptionsEditorTabs,
+  TimeSeriesQueryEditor,
+  TimeSeriesQueryEditorProps,
+} from '@perses-dev/plugin-system';
+import { produce } from 'immer';
 import { StatChartOptions } from './stat-chart-model';
 
 export type StatChartOptionsEditorProps = OptionsEditorProps<StatChartOptions>;
 
+/**
+ * Component for visually editing a Stat Chart's spec.
+ */
 export function StatChartOptionsEditor(props: StatChartOptionsEditorProps) {
-  // TODO: replace with form controls for visual editing
-  return <JSONEditor {...props} />;
+  const { onChange, value } = props;
+  const { query } = value;
+
+  const handleQueryChange: TimeSeriesQueryEditorProps['onChange'] = (next) => {
+    onChange(
+      produce(value, (draft) => {
+        draft.query = next;
+      })
+    );
+  };
+
+  return (
+    <OptionsEditorTabs
+      tabs={{
+        query: {
+          content: <TimeSeriesQueryEditor value={query} onChange={handleQueryChange} />,
+        },
+        json: {
+          content: <JSONEditor {...props} />,
+        },
+      }}
+    />
+  );
 }