+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [ENHANCEMENT] User always sees edit icons in edit mode (instead of only seeing on hover) #748
 - [ENHANCEMENT] Add unit selector to time series chart options #744
 - [BUGFIX] Fix the default config file used in the docker images and in the archive #745
+- [ENHANCEMENT] Add UI for configuring gauge panel options #761
 
 ## 0.15.0 / 2022-11-09